File restructure, deletePackage unit tests
