+++ package.json
@@ -8,18 +8,19 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "electron": "^13.1.9",
     "electron-is-dev": "^2.0.0",
     "electron-updater": "^4.3.9",
     "extract-zip": "^2.0.1",
+    "fs-extra": "^10.0.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
     "sharp": "^0.29.0",
     "styled-components": "^5.3.0"
   },
-  "main": "public/electron.js",
+  "main": "src/services/electron.js",
   "homepage": "./",
   "scripts": {
     "start": "react-scripts start",
@@ -46,12 +47,14 @@
   "devDependencies": {
     "concurrently": "^6.2.1",
     "cross-env": "^7.0.3",
-    "electron": "^13.1.9",
     "electron-builder": "^22.11.7",
     "eslint": "^7.32.0",
     "eslint-plugin-jest-dom": "^3.9.0",
+    "eslint-plugin-mocha": "^9.0.0",
     "eslint-plugin-react": "^7.24.0",
     "eslint-plugin-testing-library": "^4.12.0",
+    "react-scripts": "4.0.3",
+    "spectron": "^15.0.0",
     "wait-on": "^6.0.0"
   },
   "repository": {
