--- public/MainApi/restartApp/restartApp.js
@@ -1,10 +1,0 @@
-const { ipcMain } = require("electron");
-const { autoUpdater } = require('electron-updater');
-
-module.exports = {
-    configure: () => {
-        ipcMain.handle("restartApp", () => {
-            autoUpdater.quitAndInstall();
-        });
-    }
-};
