+++ src/services/__mocks__/electron.js
@@ -1,0 +1,6 @@
+export const ipcMain = {
+    handle: jest.fn()
+};
+export const app = {
+    getPath: jest.fn()
+}
